+++ src/briefy/leica/models/professional/workflows.py
@@ -1,6 +1,6 @@
 """Professional workflow."""
 from briefy.common.vocabularies.roles import Groups as G
-from briefy.common.vocabularies.roles import LocalRoles as LR
+from briefy.common.vocabularies.roles import LocalRolesChoices as LR
 from briefy.common.workflow import BriefyWorkflow
 from briefy.common.workflow import Permission
 from briefy.common.workflow import WorkflowState as WS