+++ src/briefy/leica/models/sync.py
@@ -4,7 +4,6 @@ job_fields_map = [
     ('professional_id', 'responsible_photographer'),
     ('qa_manager', 'qa_manager'),
     ('scout_manager', 'scouting_manager'),
-    ('finance_manager', 'finance_manager'),
     ('number_of_assets', 'number_of_photos'),
     ('job_id', 'internal_job_id'),
     ('customer_job_id', 'job_id'),